+++ src/backend/joanie/tests/core/test_api_enrollment.py
@@ -2037,7 +2037,7 @@ class EnrollmentApiTest(BaseAPITestCase):
             payment_schedule=[
                 {
                     "amount": "200.00",
-                    "due_date": "2024-01-17T00:00:00+00:00",
+                    "due_date": "2024-01-17",
                     "state": enums.PAYMENT_STATE_PAID,
                 },
             ],