+++ src/backend/joanie/tests/core/test_models_order.py
@@ -371,7 +371,7 @@ class OrderModelsTestCase(TestCase, BaseLogMixinTestCase):
             payment_schedule=[
                 {
                     "amount": "200.00",
-                    "due_date": "2024-01-17T00:00:00+00:00",
+                    "due_date": "2024-01-17",
                     "state": enums.PAYMENT_STATE_PAID,
                 }
             ],
@@ -828,7 +828,7 @@ class OrderModelsTestCase(TestCase, BaseLogMixinTestCase):
             payment_schedule=[
                 {
                     "amount": "200.00",
-                    "due_date": "2024-01-17T00:00:00+00:00",
+                    "due_date": "2024-01-17",
                     "state": enums.PAYMENT_STATE_PAID,
                 }
             ],
@@ -996,7 +996,7 @@ class OrderModelsTestCase(TestCase, BaseLogMixinTestCase):
             payment_schedule=[
                 {
                     "amount": "200.00",
-                    "due_date": "2024-01-17T00:00:00+00:00",
+                    "due_date": "2024-01-17",
                     "state": enums.PAYMENT_STATE_PAID,
                 }
             ],